corrected test data
The last build failed because, in generating the new `vendors.json` fixture, new primary keys were created; in particular, the primary key for the test vendor in the `ContractsTest` case changed from 123 to 128. Changing the `contracts.json` fixture to account for this immediately caused the test case to pass.

I don't foresee this causing any trouble down the road, seeing as the `load_fpds` command used to fetch contract data relies on the `duns` attribute of the `Vendor` model.